+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### 0.18.0
  - remove `alpha` keyword from all statistical functions. 
  - introduce new statistical tests between univariate datasets: `survival_difference_at_fixed_point_in_time_test`,...
+ - new warning message when lifelines detects possible non-unique solutions to maximum likelihood. 
+ - catch `LinAlgError: Matrix is singular.` and report back to the user. 
+ - Generally: clean up lifelines exception handling. 
  
 
 ### 0.17.5